One line added.
I try to follow 'Useful tips for a better commit message'. It says tha I shohuld describe that commit shortly one the first line (<50 symbols), and write little description about what I try to fixed and why that problem was caused. So If I will see that a huge commit message in log than that tip is not so useful. But otherwise I will do this in real projects.